+++ CHANGELOG.md
@@ -6,7 +6,8 @@
   * For analytics, send `experiment` as a parameter to `getRecommendedPaymentMethods` method
   * For analytics, send `experiment` and `paymentMethodsDisplayed` analytic metrics to FPTI via the button presented event methods
 * Venmo
-  * Add `VenmoClient` constructor with `appLinkReturnUri` to use App Links when redirecting back from the Venmo flow
+  * Add `VenmoClient` constructor with `appLinkReturnUri` argument to use App Links when redirecting back from the Venmo flow
+  * Deprecate `VenmoClient` constructor with `returnUrlScheme` argument
 
 ## 5.1.0 (2024-10-15)